+++ lib/nerves_network/dhcp_manager.ex
@@ -5,11 +5,10 @@ defmodule Nerves.Network.DHCPManager do
   require Logger
   import Nerves.Network.Utils
   use Nerves.Network.Debug
+  alias Nerves.Network.Types
 
   @moduledoc false
 
-  # The current state machine state is called "context" to avoid confusion between server
-  # state and state machine state.
   defstruct context: :removed,
             ifname: nil,
             settings: nil,
@@ -17,30 +16,30 @@ defmodule Nerves.Network.DHCPManager do
             dhcp_retry_interval: 60_000,
             dhcp_retry_timer: nil
 
+  @typedoc "Settings for starting the server."
+  @type dhcp_settings :: Nerves.Network.setup_settings
+
+  @typep context :: Types.interface_context | :dhcp | :dhcp_retry
+
+  @typedoc """
+  The current state machine state is called "context" to avoid confusion between server
+  state and state machine state.
+  """
+  @type t :: %__MODULE__{
+    context: context,
+    ifname: Types.ifname | nil,
+    settings: dhcp_settings,
+    dhcp_pid: GenServer.server() | nil,
+    dhcp_retry_interval: integer,
+    dhcp_retry_timer: reference
+  }
+
+  @doc false
+  @spec start_link(Types.ifname, dhcp_settings, GenServer.options) :: GenServer.on_start
   def start_link(ifname, settings, opts \\ []) do
     Logger.debug fn -> "DHCPManager starting.... ifname: #{inspect ifname}; settings: #{inspect settings}" end
     GenServer.start_link(__MODULE__, {ifname, settings}, opts)
   end
-
-  # defmodule EventHandler do
-  #   use GenServer
-  #
-  #   @moduledoc false
-  #
-  #   def start_link(opts) do
-  #     GenServer.start_link(__MODULE__, opts)
-  #   end
-  #
-  #   def init({manager, ifname}) do
-  #     {:ok, %{manager: manager, ifname: ifname}}
-  #   end
-  #
-  #
-  # end
-
-  #  def handle_call(a, b, c) do
-    #  Logger.debug fn -> "#{__MODULE__}: handle_call: (a=#{inspect a}, b=#{inspect b}, c=#{inspect c})" end
-    #  end
 
   def init({ifname, settings}) do
     debug_init(@debug?)
@@ -64,117 +63,135 @@ defmodule Nerves.Network.DHCPManager do
     {:ok, state}
   end
 
-  def handle_event({Nerves.NetworkInterface, :ifadded, %{ifname: ifname}}) do
-    Logger.debug "DHCPManager.EventHandler(#{ifname}) ifadded"
+  @spec handle_registry_event({Nerves.NetworkInterface, atom, %{ifname: Types.ifname}}) :: Types.ifevent
+  defp handle_registry_event({Nerves.NetworkInterface, :ifadded, %{ifname: ifname}}) do
+    Logger.debug "DHCPManager(#{ifname}) network_interface ifadded"
     :ifadded
   end
+
   # :ifmoved occurs on systems that assign stable names to removable
   # interfaces. I.e. the interface is added under the dynamically chosen
   # name and then quickly renamed to something that is stable across boots.
-  def handle_event({Nerves.NetworkInterface, :ifmoved, %{ifname: ifname}}) do
-    Logger.debug "DHCPManager.EventHandler(#{ifname}) ifadded (moved)"
+  defp handle_registry_event({Nerves.NetworkInterface, :ifmoved, %{ifname: ifname}}) do
+    Logger.debug "DHCPManager(#{ifname}) network_interface ifadded (moved)"
     :ifadded
   end
-  def handle_event({Nerves.NetworkInterface, :ifremoved, %{ifname: ifname}}) do
-    Logger.debug "DHCPManager.EventHandler(#{ifname}) ifremoved"
+
+  defp handle_registry_event({Nerves.NetworkInterface, :ifremoved, %{ifname: ifname}}) do
+    Logger.debug "DHCPManager(#{ifname}) network_interface ifremoved"
     :ifremoved
   end
 
   # Filter out ifup and ifdown events
   # :is_up reports whether the interface is enabled or disabled (like by the wifi kill switch)
   # :is_lower_up reports whether the interface as associated with an AP
-  def handle_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: true}}) do
-    Logger.debug "DHCPManager.EventHandler(#{ifname}) ifup"
+  defp handle_registry_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: true}}) do
+    Logger.debug "DHCPManager(#{ifname}) network_interface ifup"
     :ifup
   end
-  def handle_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: false}}) do
-    Logger.debug "DHCPManager.EventHandler(#{ifname}) ifdown"
+
+  defp handle_registry_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: false}}) do
+    Logger.debug "DHCPManager(#{ifname}) network_interface ifdown"
     :ifdown
   end
 
-  # # DHCP events
-  # # :bound, :renew, :deconfig, :nak
-
-
-  def handle_event({Nerves.NetworkInterface, event, %{ifname: ifname}}) do
-    Logger.debug "DHCPManager.EventHandler(#{ifname}): ignoring event: #{inspect event}"
+  defp handle_registry_event({Nerves.NetworkInterface, event, %{ifname: ifname}}) do
+    Logger.debug "DHCPManager(#{ifname}): ignoring event: #{inspect event}"
     :noop
   end
 
+  defp handle_registry_event({Nerves.Udhcpc, :bound, %{ifname: ifname} = info}) do
+    Logger.info "WiFiManager(#{ifname}): udhcp bound"
+    {:bound, info}
+  end
+
+  defp handle_registry_event({Nerves.Udhcpc, :deconfig, %{ifname: ifname} = info}) do
+    Logger.info "WiFiManager(#{ifname}): udhcp deconfig"
+    {:deconfig, info}
+  end
+
+  defp handle_registry_event({Nerves.Udhcpc, :nak, %{ifname: ifname} = info}) do
+    Logger.info "WiFiManager(#{ifname}): udhcp nak"
+    {:nak, info}
+  end
+
+  defp handle_registry_event({Nerves.Udhcpc, :renew, %{ifname: ifname} = info}) do
+    Logger.info "WiFiManager(#{ifname}): udhcp renew"
+    {:renew, info}
+  end
+
+  defp handle_registry_event({Nerves.Udhcpc, event, %{ifname: ifname}}) do
+    Logger.info "WiFiManager(#{ifname}): ignoring event: #{inspect event}"
+    :noop
+  end
+
+  # Handle Network Interface events coming in from SystemRegistry.
   def handle_info({Nerves.NetworkInterface, _, ifstate} = event, %{ifname: ifname} = s) do
-    event = handle_event(event)
+    event = handle_registry_event(event)
     scope(ifname) |> SystemRegistry.update(ifstate)
     s = consume(s.context, event, s)
     Logger.debug "DHCPManager(#{s.ifname}, #{s.context}) got event #{inspect event}"
     {:noreply, s}
   end
 
+  # Handle Udhcpc events coming from SystemRegistry.
   def handle_info({Nerves.Udhcpc, event, info}, %{ifname: ifname} = s) do
-    Logger.debug "DHCPManager.EventHandler(#{s.ifname}) udhcpc #{inspect event}"
+    Logger.debug "DHCPManager(#{s.ifname}) udhcpc #{inspect event}"
     scope(ifname) |> SystemRegistry.update(info)
     s = consume(s.context, {event, info}, s)
     {:noreply, s}
   end
 
+  # Comes from the timer.
   def handle_info(:dhcp_retry, s) do
     s = consume(s.context, :dhcp_retry, s)
     {:noreply, s}
   end
 
+  # Catch all.
   def handle_info(event, s) do
-    Logger.debug "DHCPManager.EventHandler(#{s.ifname}): ignoring event: #{inspect event}"
+    Logger.debug "DHCPManager(#{s.ifname}): ignoring event: #{inspect event}"
     {:noreply, s}
   end
 
   ## State machine implementation
+  @spec goto_context(t, context) :: t
   defp goto_context(state, newcontext) do
     %Nerves.Network.DHCPManager{state | context: newcontext}
   end
 
+  @typedoc "Event for the state machine."
+  @type event :: Types.ifevent | Nerves.Network.Udhcpc.event
+
+  @spec consume(context, event, t) :: t
   defp consume(_, :noop, state), do: state
+
   ## Context: :removed
   defp consume(:removed, :ifadded, state) do
-    case Nerves.NetworkInterface.ifup(state.ifname) do
-      :ok ->
-        {:ok, status} = Nerves.NetworkInterface.status state.ifname
-        notify(Nerves.NetworkInterface, state.ifname, :ifchanged, status)
+    :ok = Nerves.NetworkInterface.ifup(state.ifname)
 
-        state
-          |> goto_context(:down)
-      {:error, _} ->
-        # The interface isn't quite up yet. Retry
-        Process.send_after self(), :retry_ifadded, 250
-        state
-          |> goto_context(:retry_add)
-    end
-  end
-  defp consume(:removed, :retry_ifadded, state), do: state
-  defp consume(:removed, :ifdown, state), do: state
-
-  ## Context: :retry_add
-  defp consume(:retry_add, :ifremoved, state) do
-    state
-      |> goto_context(:removed)
-  end
-  defp consume(:retry_add, :retry_ifadded, state) do
-    {:ok, status} = Nerves.NetworkInterface.status(state.ifname)
+    {:ok, status} = Nerves.NetworkInterface.status state.ifname
     notify(Nerves.NetworkInterface, state.ifname, :ifchanged, status)
 
-    state
-      |> goto_context(:down)
+    state |> goto_context(:down)
   end
+
+  defp consume(:removed, :ifdown, state), do: state
 
   ## Context: :down
   defp consume(:down, :ifadded, state), do: state
+
   defp consume(:down, :ifup, state) do
     state
       |> start_udhcpc
       |> goto_context(:dhcp)
   end
+
   defp consume(:down, :ifdown, state) do
     state
       |> stop_udhcpc
   end
+
   defp consume(:down, :ifremoved, state) do
     state
       |> stop_udhcpc
@@ -183,20 +200,22 @@ defmodule Nerves.Network.DHCPManager do
 
   ## Context: :dhcp
   defp consume(:dhcp, :ifup, state), do: state
+
   defp consume(:dhcp, {:deconfig, _info}, state), do: state
+
   defp consume(:dhcp, {:bound, info}, state) do
     state
       |> configure(info)
       |> goto_context(:up)
   end
+
   defp consume(:dhcp, {:leasefail, _info}, state) do
     dhcp_retry_timer = Process.send_after(self(), :dhcp_retry, state.dhcp_retry_interval)
     %{state | dhcp_retry_timer: dhcp_retry_timer}
       |> stop_udhcpc
-      |> start_link_local
       |> goto_context(:up)
-
   end
+
   defp consume(:dhcp, :ifdown, state) do
     state
       |> stop_udhcpc
@@ -236,6 +255,7 @@ defmodule Nerves.Network.DHCPManager do
     state
   end
 
+  @spec stop_udhcpc(t) :: t
   defp stop_udhcpc(state) do
     if is_pid(state.dhcp_pid) do
       Nerves.Network.Udhcpc.stop(state.dhcp_pid)
@@ -244,27 +264,22 @@ defmodule Nerves.Network.DHCPManager do
       state
     end
   end
+
+  @spec start_udhcpc(t) :: t
   defp start_udhcpc(state) do
     state = stop_udhcpc(state)
     {:ok, pid} = Nerves.Network.Udhcpc.start_link(state.ifname)
     %Nerves.Network.DHCPManager{state | dhcp_pid: pid}
   end
 
-  defp start_link_local(state) do
-    {:ok, ifsettings} = Nerves.NetworkInterface.status(state.ifname)
-    ip = generate_link_local(ifsettings.mac_address)
-    scope(state.ifname)
-    |> SystemRegistry.update(%{ipv4_address: ip})
-    :ok = Nerves.NetworkInterface.setup(state.ifname, [ipv4_address: ip])
-    state
-  end
-
+  @spec configure(t, Types.udhcp_info) :: t
   defp configure(state, info) do
     :ok = Nerves.NetworkInterface.setup(state.ifname, info)
     :ok = Nerves.Network.Resolvconf.setup(Nerves.Network.Resolvconf, state.ifname, info)
     state
   end
 
+  @spec deconfigure(t) :: t
   defp deconfigure(state) do
     :ok = Nerves.Network.Resolvconf.clear(Nerves.Network.Resolvconf, state.ifname)
     state