+++ lib/nerves_network/linklocal_manager.ex
@@ -2,15 +2,24 @@ defmodule Nerves.Network.LinkLocalManager do
   use GenServer
   require Logger
   import Nerves.Network.Utils
+  alias Nerves.Network.Types
 
   @moduledoc false
 
-  # The current state machine state is called "context" to avoid confusion between server
-  # state and state machine state.
   defstruct context: :removed,
             ifname: nil,
             settings: nil
 
+  @typep context :: Types.interface_context
+
+  @type t :: %__MODULE__{
+    context: context,
+    ifname: Types.ifname | nil,
+    settings: Nerves.Network.setup_settings | nil
+  }
+
+  @doc false
+  @spec start_link(Types.ifname, Nerves.Network.setup_settings, GenServer.options) :: GenServer.on_start
   def start_link(ifname, settings, opts \\ []) do
     GenServer.start_link(__MODULE__, {ifname, settings}, opts)
   end
@@ -33,41 +42,42 @@ defmodule Nerves.Network.LinkLocalManager do
     {:ok, state}
   end
 
-  def handle_event({Nerves.NetworkInterface, :ifadded, %{ifname: ifname}}) do
-    Logger.debug "LinkLocalManager.EventHandler(#{ifname}) ifadded"
+  @spec handle_registry_event({Nerves.NetworkInterface, atom, %{ifname: Types.ifname, is_lower_up: boolean}}) :: Types.ifevent
+  defp handle_registry_event({Nerves.NetworkInterface, :ifadded, %{ifname: ifname}}) do
+    Logger.debug "LinkLocalManager(#{ifname}) network_interface ifadded"
     :ifadded
   end
   # :ifmoved occurs on systems that assign stable names to removable
   # interfaces. I.e. the interface is added under the dynamically chosen
   # name and then quickly renamed to something that is stable across boots.
-  def handle_event({Nerves.NetworkInterface, :ifmoved, %{ifname: ifname}}) do
-    Logger.debug "LinkLocalManager.EventHandler(#{ifname}) ifadded (moved)"
+  defp handle_registry_event({Nerves.NetworkInterface, :ifmoved, %{ifname: ifname}}) do
+    Logger.debug "LinkLocalManager(#{ifname}) network_interface ifadded (moved)"
     :ifadded
   end
-  def handle_event({Nerves.NetworkInterface, :ifremoved, %{ifname: ifname}}) do
-    Logger.debug "LinkLocalManager.EventHandler(#{ifname}) ifremoved"
+  defp handle_registry_event({Nerves.NetworkInterface, :ifremoved, %{ifname: ifname}}) do
+    Logger.debug "LinkLocalManager(#{ifname}) network_interface ifremoved"
     :ifremoved
   end
 
   # Filter out ifup and ifdown events
   # :is_up reports whether the interface is enabled or disabled (like by the wifi kill switch)
   # :is_lower_up reports whether the interface as associated with an AP
-  def handle_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: true}}) do
-    Logger.debug "LinkLocalManager.EventHandler(#{ifname}) ifup"
+  defp handle_registry_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: true}}) do
+    Logger.debug "LinkLocalManager(#{ifname}) network_interface ifup"
     :ifup
   end
-  def handle_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: false}}) do
-    Logger.debug "LinkLocalManager.EventHandler(#{ifname}) ifdown"
+  defp handle_registry_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: false}}) do
+    Logger.debug "LinkLocalManager(#{ifname}) network_interface ifdown"
     :ifdown
   end
 
-  def handle_event({Nerves.NetworkInterface, event, %{ifname: ifname}}) do
-    Logger.debug "LinkLocalManager.EventHandler(#{ifname}): ignoring event: #{inspect event}"
+  defp handle_registry_event({Nerves.NetworkInterface, event, %{ifname: ifname}}) do
+    Logger.debug "LinkLocalManager(#{ifname}): ignoring event: #{inspect event}"
     :noop
   end
 
   def handle_info({Nerves.NetworkInterface, _, ifstate} = event, %{ifname: ifname} = s) do
-    event = handle_event(event)
+    event = handle_registry_event(event)
     scope(ifname) |> SystemRegistry.update(ifstate)
     s = consume(s.context, event, s)
     Logger.debug "LinkLocalManager(#{s.ifname}, #{s.context}) got event #{inspect event}"
@@ -75,45 +85,32 @@ defmodule Nerves.Network.LinkLocalManager do
   end
 
   def handle_info(event, s) do
-    Logger.debug "LinkLocalManager.EventHandler(#{s.ifname}): ignoring event: #{inspect event}"
+    Logger.debug "LinkLocalManager(#{s.ifname}): ignoring event: #{inspect event}"
     {:noreply, s}
   end
 
+  @type event :: Types.ifevent
+
   ## State machine implementation
+  @spec goto_context(t, context) :: t
   defp goto_context(state, newcontext) do
     %Nerves.Network.LinkLocalManager{state | context: newcontext}
   end
 
+  @spec consume(context, event, t) :: t
   defp consume(_, :noop, state), do: state
+
   ## Context: :removed
   defp consume(:removed, :ifadded, state) do
-    case Nerves.NetworkInterface.ifup(state.ifname) do
-      :ok ->
-        {:ok, status} = Nerves.NetworkInterface.status(state.ifname)
-        notify(Nerves.NetworkInterface, state.ifname, :ifchanged, status)
-
-        goto_context(state, :down)
-      {:error, _} ->
-        # The interface isn't quite up yet. Retry
-        Process.send_after self(), :retry_ifadded, 250
-        goto_context(state, :retry_add)
-    end
-  end
-  defp consume(:removed, :retry_ifadded, state), do: state
-  defp consume(:removed, :ifdown, state), do: state
-  defp consume(:removed, :ifremoved, state), do: state
-
-  ## Context: :retry_add
-  defp consume(:retry_add, :ifadded, state), do: state
-  defp consume(:retry_add, :retry_ifadded, state) do
+    :ok = Nerves.NetworkInterface.ifup(state.ifname)
     {:ok, status} = Nerves.NetworkInterface.status(state.ifname)
     notify(Nerves.NetworkInterface, state.ifname, :ifchanged, status)
 
     goto_context(state, :down)
   end
-  defp consume(:retry_add, :ifremoved, state) do
-    goto_context(state, :removed)
-  end
+
+  defp consume(:removed, :ifdown, state), do: state
+  defp consume(:removed, :ifremoved, state), do: state
 
   ## Context: :down
   defp consume(:down, :ifadded, state), do: state
@@ -134,11 +131,12 @@ defmodule Nerves.Network.LinkLocalManager do
     goto_context(state, :down)
   end
 
+  @spec start_link_local(t) :: t
   defp start_link_local(state) do
     {:ok, ifsettings} = Nerves.NetworkInterface.status(state.ifname)
     ip = generate_link_local(ifsettings.mac_address)
     scope(state.ifname)
-    |> SystemRegistry.update(%{ipv4_address: ip})
+      |> SystemRegistry.update(%{ipv4_address: ip})
     :ok = Nerves.NetworkInterface.setup(state.ifname, [ipv4_address: ip])
     state
   end