+++ lib/nerves_network/static_manager.ex
@@ -1,12 +1,26 @@
 defmodule Nerves.Network.StaticManager do
+  @moduledoc false
+
   use GenServer
   import Nerves.Network.Utils
+  alias Nerves.Network.Types
   require Logger
 
   defstruct context: :removed,
             ifname: nil,
             settings: nil
 
+  @typep context :: Types.interface_context
+
+  @typedoc "GenServer state."
+  @type t :: %__MODULE__{
+    context: context,
+    ifname: Types.ifname | nil,
+    settings: Nerves.Network.setup_settings | nil
+  }
+
+  @doc false
+  @spec start_link(Types.ifname, Nerves.Network.setup_settings, GenServer.options) :: GenServer.on_start
   def start_link(ifname, settings, opts \\ []) do
     GenServer.start_link(__MODULE__, {ifname, settings}, opts)
   end
@@ -33,81 +47,66 @@ defmodule Nerves.Network.StaticManager do
   end
 
   def handle_info({Nerves.NetworkInterface, _, ifstate} = event, %{ifname: ifname} = s) do
-    event = handle_event(event)
+    event = handle_registry_event(event)
     scope(ifname) |> SystemRegistry.update(ifstate)
     s = consume(s.context, event, s)
     Logger.info "StaticManager(#{s.ifname}, #{s.context}) got event #{inspect event}"
     {:noreply, s}
   end
 
-  def handle_event({Nerves.NetworkInterface, :ifadded, %{ifname: ifname}}) do
-    Logger.info "StaticManager.EventHandler(#{ifname}) ifadded"
+  @spec handle_registry_event({Nerves.NetworkInterface, atom, %{ifname: Types.ifname, is_lower_up: boolean}}) :: Types.ifevent
+  defp handle_registry_event({Nerves.NetworkInterface, :ifadded, %{ifname: ifname}}) do
+    Logger.info "StaticManager(#{ifname}) network_interface ifadded"
     :ifadded
   end
   # :ifmoved occurs on systems that assign stable names to removable
   # interfaces. I.e. the interface is added under the dynamically chosen
   # name and then quickly renamed to something that is stable across boots.
-  def handle_event({Nerves.NetworkInterface, :ifmoved, %{ifname: ifname}}) do
-    Logger.info "StaticManager.EventHandler(#{ifname}) ifadded (moved)"
+  defp handle_registry_event({Nerves.NetworkInterface, :ifmoved, %{ifname: ifname}}) do
+    Logger.info "StaticManager(#{ifname}) network_interface ifadded (moved)"
     :ifadded
   end
-  def handle_event({Nerves.NetworkInterface, :ifremoved, %{ifname: ifname}}) do
-    Logger.info "StaticManager.EventHandler(#{ifname}) ifremoved"
+  defp handle_registry_event({Nerves.NetworkInterface, :ifremoved, %{ifname: ifname}}) do
+    Logger.info "StaticManager(#{ifname}) network_interface ifremoved"
     :ifremoved
   end
 
   # Filter out ifup and ifdown events
   # :is_up reports whether the interface is enabled or disabled (like by the wifi kill switch)
   # :is_lower_up reports whether the interface as associated with an AP
-  def handle_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: true}}) do
-    Logger.info "StaticManager.EventHandler(#{ifname}) ifup"
+  defp handle_registry_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: true}}) do
+    Logger.info "StaticManager(#{ifname}) network_interface ifup"
     :ifup
   end
-  def handle_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: false}}) do
-    Logger.info "StaticManager.EventHandler(#{ifname}) ifdown"
+  defp handle_registry_event({Nerves.NetworkInterface, :ifchanged, %{ifname: ifname, is_lower_up: false}}) do
+    Logger.info "StaticManager(#{ifname}) network_interface ifdown"
     :ifdown
   end
 
-  def handle_event({Nerves.NetworkInterface, event, %{ifname: ifname}}) do
-    Logger.info "StaticManager.EventHandler(#{ifname}): ignoring event: #{inspect event}"
+  defp handle_registry_event({Nerves.NetworkInterface, event, %{ifname: ifname}}) do
+    Logger.info "StaticManager(#{ifname}): network_interface ignoring event: #{inspect event}"
     :noop
   end
 
   ## State machine implementation
+  @spec goto_context(t, context) :: t
   defp goto_context(state, newcontext) do
     %Nerves.Network.StaticManager{state | context: newcontext}
   end
 
+  @typedoc "Network event to be consumed."
+  @type event :: Types.ifevent
+
+  @spec consume(context, event, t) :: t
   defp consume(_, :noop, state), do: state
 
+  ## Context: removed
   defp consume(:removed, :ifadded, state) do
-    case Nerves.NetworkInterface.ifup(state.ifname) do
-      :ok ->
-        {:ok, status} = Nerves.NetworkInterface.status state.ifname
-        notify(Nerves.NetworkInterface, state.ifname, :ifchanged, status)
-
-        state
-          |> goto_context(:down)
-      {:error, _} ->
-        # The interface isn't quite up yet. Retry
-        Process.send_after self(), :retry_ifadded, 250
-        state
-          |> goto_context(:retry_add)
-    end
-  end
-  defp consume(:removed, :retry_ifadded, state), do: state
-
-  ## Context: :retry_add
-  defp consume(:retry_add, :ifremoved, state) do
-    state
-      |> goto_context(:removed)
-  end
-  defp consume(:retry_add, :retry_ifadded, state) do
-    {:ok, status} = Nerves.NetworkInterface.status(state.ifname)
+    :ok = Nerves.NetworkInterface.ifup(state.ifname)
+    {:ok, status} = Nerves.NetworkInterface.status state.ifname
     notify(Nerves.NetworkInterface, state.ifname, :ifchanged, status)
 
-    state
-      |> goto_context(:down)
+    state |> goto_context(:down)
   end
 
   ## Context: :down
@@ -116,25 +115,33 @@ defmodule Nerves.Network.StaticManager do
       |> configure
       |> goto_context(:up)
   end
+
   defp consume(:down, :ifdown, state), do: state
+
   defp consume(:down, :ifremoved, state) do
     state
       |> goto_context(:removed)
   end
 
+  ## Context: :up
   defp consume(:up, :ifup, state), do: state
+
   defp consume(:up, :ifdown, state) do
     state
       |> deconfigure
       |> goto_context(:down)
   end
 
+  defp consume(:up, :ifadded, state), do: state
+
+  @spec configure(t) :: t
   defp configure(state) do
     :ok = Nerves.NetworkInterface.setup(state.ifname, state.settings)
     :ok = Nerves.Network.Resolvconf.setup(Nerves.Network.Resolvconf, state.ifname, state.settings)
     state
   end
 
+  @spec deconfigure(t) :: t
   defp deconfigure(state) do
     :ok = Nerves.Network.Resolvconf.clear(Nerves.Network.Resolvconf, state.ifname)
     state